[esp32-wifi-accessory] Extract reset button logic from GarageRemote class
