+++ esp32-wifi-accessory/main/CMakeLists.txt
@@ -1,1 +1,1 @@
-idf_component_register(SRCS garage_remote.cpp log_config.c main.cpp wifi.c)
+idf_component_register(SRCS garage_remote.cpp homekit.c log_config.c main.cpp wifi.c)
