+++ esp32-wifi-accessory/main/homekit.h
@@ -1,0 +1,8 @@
+#ifndef IPAD_CAR_INTEGRATION_HOMEKIT_H_
+#define IPAD_CAR_INTEGRATION_HOMEKIT_H_
+
+#include <driver/gpio.h>
+
+void initializeHomeKitResetButton(gpio_num_t resetButtonPin);
+
+#endif
